+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 
+-   Write "Unknown cartesian CRS" when saving gdf without a CRS to GPKG (#368).
 -   `read_arrow` and `open_arrow` now provide
     [GeoArrow-compliant extension metadata](https://geoarrow.org/extension-types.html),
     including the CRS, when using GDAL 3.8 or higher (#366).